+++ lamb/middleware/execution_time.py
@@ -31,23 +31,57 @@ class LambExecutionTimeMiddleware(MiddlewareMixin):
         with contextlib.suppress(Exception):
             request.lamb_execution_meter.append_marker(message)
 
+    # settings: memoize
     @lazy_default_ro(default=[])
-    def settings_skip_methods(self) -> list[str]:
+    def _settings_skip_methods(self) -> list[str]:
         result = dpath_value(settings, "LAMB_EXECUTION_TIME_SKIP_METHODS", str, transform=tf_list_string, default=[])
         result = [r.upper() for r in result]
         logger.debug(f"<{self.__class__.__name__}>. settings_skip_methods: {result}")
         return result
 
     @lazy_default_ro(default=False)
-    def settings_should_store(self) -> bool:
+    def _settings_should_store(self) -> bool:
         result = dpath_value(settings, "LAMB_EXECUTION_TIME_STORE", str, transform=transform_boolean)
         logger.debug(f"<{self.__class__.__name__}>. settings_should_store: {result}")
         return result
 
     @lazy_default_ro(default={})
-    def settings_store_rates(self) -> dict[tuple[str, str], float]:
+    def _settings_store_rates(self) -> dict[tuple[str, str], float]:
         result = settings.LAMB_EXECUTION_TIME_STORE_RATES
         logger.debug(f"<{self.__class__.__name__}>. settings_store_rates: {result}")
+        return result
+
+    @lazy_default_ro(default=None)
+    def _settings_log_total_level(self) -> int | None:
+        result = settings.LAMB_EXECUTION_TIME_LOG_TOTAL_LEVEL
+        if isinstance(result, str):
+            result = logging.getLevelName(result.upper())
+        elif isinstance(result, int):
+            pass
+        elif result is None:
+            return None
+        else:
+            logger.warning(f"could not determine LAMB_EXECUTION_TIME_LOG_TOTAL_LEVEL value: {result}")
+            raise ValueError
+
+        logger.critical(f"<{self.__class__.__name__}>. settings_log_total_level: {result}")
+        return result
+
+    @lazy_default_ro(default=None)
+    def _settings_log_markers_level(self) -> int | None:
+        logger.critical("looking for: _settings_log_markers_level")
+        result = settings.LAMB_EXECUTION_TIME_LOG_MARKERS_LEVEL
+        if isinstance(result, str):
+            result = logging.getLevelName(result.upper())
+        elif isinstance(result, int):
+            pass
+        elif result is None:
+            return None
+        else:
+            logger.warning(f"could not determine LAMB_EXECUTION_TIME_LOG_MARKERS_LEVEL value: {result}")
+            raise ValueError
+
+        logger.critical(f"<{self.__class__.__name__}>. settings_log_markers_level: {result}")
         return result
 
     # utils
@@ -56,7 +90,8 @@ class LambExecutionTimeMiddleware(MiddlewareMixin):
         request.lamb_execution_meter = ExecutionTimeMeter()
 
     def _finish(self, request: LambRequest, response: HttpResponse | None, exception: Exception | None):
-        """Stores collected data in database"""
+        """Stores collected data in database and logs"""
+        # prepare base container and record
         metric = LambExecutionTimeMetric()
         metric.http_method = request.method
         metric.headers = dict(request.headers)
@@ -64,7 +99,16 @@ class LambExecutionTimeMiddleware(MiddlewareMixin):
         metric.device_info = request.lamb_device_info
         metric.status_code = response.status_code if response else None
 
-        # get context and execution time
+        # append app_name and url_name
+        try:
+            resolved = resolve(request.path)
+            metric.app_name = resolved.app_name
+            metric.url_name = resolved.url_name
+        except Exception:
+            pass
+
+        # finalize meter, collect markers and append context
+        time_measure = None
         try:
             time_measure = request.lamb_execution_meter
 
@@ -92,16 +136,8 @@ class LambExecutionTimeMiddleware(MiddlewareMixin):
         except Exception:
             pass
 
-        # get app name and url name
-        try:
-            resolved = resolve(request.path)
-            metric.app_name = resolved.app_name
-            metric.url_name = resolved.url_name
-        except Exception:
-            pass
-
-        # store
-        if request.method not in self.settings_skip_methods and self.settings_should_store:
+        # store: database
+        if request.method not in self._settings_skip_methods and self._settings_should_store:
             try:
                 # logger.warning(f'analyze store rate: {self.store_rates}')
                 # database
@@ -113,12 +149,8 @@ class LambExecutionTimeMiddleware(MiddlewareMixin):
                 logger.error(f"<{self.__class__.__name__}>. metrics store failed: {e}")
                 pass
 
-        # log total
-        level = settings.LAMB_EXECUTION_TIME_LOG_TOTAL_LEVEL
-        if level:
-            if isinstance(level, str):
-                # TODO: fix and migrate to mapping - can produce wrong levels if not found
-                level = logging.getLevelName(level.upper())
+        # store: logging
+        if level_total := self._settings_log_total_level:
             msg = (
                 f'"{request.method} {request.get_full_path()}" {request.lamb_execution_meter.get_total_time():.6f} sec.'
             )
@@ -138,7 +170,13 @@ class LambExecutionTimeMiddleware(MiddlewareMixin):
                     "streaming": None,
                     "content_length": None,
                 }
-            logger.log(level, msg, extra=extra)
+            logger.log(level_total, msg, extra=extra)
+
+        logger.critical(f"before check: {self._settings_log_markers_level=} and {time_measure=}")
+        if level_markers := self._settings_log_markers_level:
+            if time_measure is not None:
+                for index, m in enumerate(time_measure.get_log_list()):
+                    logger.log(level_markers, f"<{self.__class__.__name__}>. [{index}] {m}")
 
     # lifecycle
     def process_request(self, request: LambRequest):